+++ ignite/templates/app/proto.go
@@ -2,17 +2,22 @@ package app
 
 import (
 	"embed"
+	"fmt"
+	"strings"
 
 	"github.com/gobuffalo/genny/v2"
+	"github.com/gobuffalo/plush/v4"
 
+	"github.com/ignite/cli/v29/ignite/pkg/xembed"
 	"github.com/ignite/cli/v29/ignite/pkg/xgenny"
+	"github.com/ignite/cli/v29/ignite/templates/field/plushhelpers"
 )
 
-//go:embed files/proto/* files/buf.work.yaml
+//go:embed files/{{protoDir}}/* files/buf.work.yaml.plush
 var fsProto embed.FS
 
 // NewBufGenerator returns the generator to buf build files.
-func NewBufGenerator(appPath string) (*genny.Generator, error) {
+func NewBufGenerator(appPath, protoDir string) (*genny.Generator, error) {
 	var (
 		g        = genny.New()
 		template = xgenny.NewEmbedWalker(
@@ -21,5 +26,33 @@ func NewBufGenerator(appPath string) (*genny.Generator, error) {
 			appPath,
 		)
 	)
-	return g, xgenny.Box(g, template)
+	if err := xgenny.Box(g, template); err != nil {
+		return nil, err
+	}
+
+	ctx := plush.NewContext()
+	ctx.Set("ProtoDir", protoDir)
+
+	plushhelpers.ExtendPlushContext(ctx)
+	g.Transformer(xgenny.Transformer(ctx))
+	g.Transformer(genny.Replace("{{protoDir}}", protoDir))
+
+	return g, nil
+}
+
+func CutTemplatePrefix(name string) (string, bool) {
+	return strings.CutPrefix(name, fmt.Sprintf("%s/", "{{protoDir}}"))
+}
+
+// BufFiles returns a list of Buf.Build files.
+func BufFiles() ([]string, error) {
+	files, err := xembed.FileList(fsProto, "files")
+	if err != nil {
+		return nil, err
+	}
+	// remove all .plush extensions.
+	for i, file := range files {
+		files[i] = strings.TrimSuffix(file, ".plush")
+	}
+	return files, nil
 }